==> pipeline/1_dedup_tab6.R <==
# Define variables to preserve
vars_to_keep <- c("pipeline_dir", "source_if_exists", "scripts_to_run")

# Remove all objects except those in vars_to_keep
rm(list = setdiff(ls(), vars_to_keep))

library(dummy)
library(lubridate)
library(tidyr)
library(dplyr)
library(tidyverse)
library(magrittr)
library(readxl)
library(readr)

source("./global_vars.R")

message("Only need to run once per batch!")

# TODO: this is to filter Table 6 for year_w and "UCH"
# TODO: Simple dedup by person id, encounter, op date
# colnames(table2)
table2 <- read_csv(fp_tab2)
table(table2$FacilityName)

# table(table2 )

table6 <- read_csv(fp_tab6)

# sort(unique(table6$PrimarySurgeonName))
# Anna Gleisner
# Kenneth HuntS
# "FacilityName" %in% colnames(table6)

data_full <- table6 %>%
  left_join(table2 %>% select(arb_person_id,
                              arb_encounter_id,
                              FacilityName,
                              DepartmentExternalName,
                              EncounterDate),
            by = c("arb_person_id", "arb_encounter_id")) %>%
  mutate(operationDate = as.Date(SurgeryDate)) 

table(data_full$FacilityName)
# %>%
#   filter(FacilityName == "UCH")

tab6_fa_path <- paste0("../processed_data/", save_date_str, "/", "table6_facility.csv")
write.csv(data_full, tab6_fa_path)
          # TODO: row.name = TRUE

data_full <- read_csv(tab6_fa_path)
data_full %<>% dplyr::rename(X = ...1)


# row names, row numbers as ...1
# data_full$...1

data_full$operationDate <- as.Date(data_full$operationDate, format = "%Y-%m-%d")
data <- data_full %>%
  filter(year(operationDate) == year_w) 
rvu <- read_excel("../data/PPRRVU23_JAN_modified.xlsx")
data$arb_person_id <- data$arb_person_id
data$arb_encounter_id <- data$arb_encounter_id
data$operationDate <- as.Date(data$operationDate, format = "%Y-%m-%d")

# glimpse(data_full)

surgeons <- sort(unique(data$PrimarySurgeonName))

all_top_rvus <- data_frame(X = c(), arb_person_id = c(), arb_encounter_id = c(),
                           HCPCS = c(), DESCRIPTION = c(),
                           CODE = c(), WORK_RVU = c(),
                           operationDate = c(), DateDiff = c(),
                           Include = c())

# TODO: What is X, column X not found in data_s, data, data_full
# column X is row number, auto generated by write.csv

for (s in c(1:length(surgeons))) {
  data_s <- data %>%
    filter(PrimarySurgeonName == surgeons[s])
  patients <- sort(unique(data_s$arb_person_id))
  
  for (i in c(1:length(patients))) {
    cpts <- data_s %>%
      filter(arb_person_id == patients[i]) %>%
      dplyr::select(X, arb_person_id, arb_encounter_id, ProcedureName,
                    CuMedCptCode, operationDate) %>%
      arrange(operationDate)
    
    top_rvus <- data_frame(X = c(), arb_person_id = c(), arb_encounter_id = c(),
                           HCPCS = c(), DESCRIPTION = c(),
                           CODE = c(), WORK_RVU = c(),
                           operationDate = c(), DateDiff = c())
    for (j in c(1:nrow(cpts))) {
      X <- cpts[j, "X"]
      # FIX by adding as.character
      tag <- cpts[j, "CuMedCptCode"] %>% as.character()
      date <- cpts[j, "operationDate"]
      arb_person_id1 <- cpts[j, "arb_person_id"]
      arb_encounter_id1 <- cpts[j, "arb_encounter_id"]
      # If `tag` is `NA`, this will produce an error.
      if (!is.na(tag) && tag != "") {
        top_rvus <- rbind(top_rvus, 
                          data_frame(X = X, arb_person_id = arb_person_id1,
                                     arb_encounter_id = arb_encounter_id1,
                                     HCPCS = strsplit(tag, "|", fixed = TRUE)[[1]]) %>%
                            mutate(is_surgery = ifelse(is.na(as.numeric(HCPCS)), 0,
                                                       ifelse((as.numeric(HCPCS) <= 69999) &
                                                                (as.numeric(HCPCS) >= 10000), 1, 0))) %>%
                            filter(is_surgery == 1) %>%
                            left_join(rvu, by = "HCPCS") %>%
                            arrange(desc(WORK_RVU)) %>%
                            slice(1) %>%
                            mutate(operationDate = date,
                                   DateDiff = 0) 
        )
        # TODO: what if tag is NA 
      } else if (is.na(tag) | tag == "") {
        top_rvus <- rbind(top_rvus, 
                          data_frame(X = X, arb_person_id = arb_person_id1,
                                     arb_encounter_id = arb_encounter_id1,
                                     HCPCS = tag,
                                     is_surgery = NA,
                                     DESCRIPTION = NA,
                                     CODE = NA,
                                     WORK_RVU = NA) %>%
                            mutate(operationDate = date,
                                   DateDiff = 0) 
        )
      }
    }
    # FIX top_rvus DD
    # missing value?
    # colnames(top_rvus)
    # top_rvus$DD[2]
top_rvus <- top_rvus %>% 
      mutate(DateDiff = operationDate - lag(operationDate, 1),
             DD = DateDiff,
             Include = ifelse(is.na(DateDiff), 1, 0)) 
# FIX: top_rvus$DD[k] `[.data.frame`(top_rvus$DD, k) : undefined columns selected
# %>% as.data.frame()
# browser()
# Flatten nested column
top_rvus$X <- top_rvus[["X"]][["X"]]
top_rvus$DD <- top_rvus[["DD"]][["operationDate"]]
top_rvus$Include <- top_rvus[["Include"]][, "operationDate"]
top_rvus$operationDate <- top_rvus[["operationDate"]][["operationDate"]]
top_rvus$arb_person_id <- top_rvus[["arb_person_id"]][["arb_person_id"]]
top_rvus$arb_encounter_id <- top_rvus[["arb_encounter_id"]][["arb_encounter_id"]]
top_rvus$DateDiff <- top_rvus[["DateDiff"]][["operationDate"]]


top_rvus <- as.data.frame(top_rvus)
    if (nrow(top_rvus) > 1) {
      for (k in c(2:nrow(top_rvus))) {
        if (top_rvus$DD[k] > 30) {
          top_rvus$Include[k] <- 1
          top_rvus$DD <- top_rvus$DD - top_rvus$DD[k]
        }
      }
    }
    top_rvus <- top_rvus %>%
      filter(Include == 1)
    all_top_rvus <- rbind(all_top_rvus, top_rvus)
  }
}


data_f <- data %>%
  filter(X %in% all_top_rvus$X) %>%
  dplyr::select(-X, -FacilityName, -EncounterDate, -operationDate)

# saveRDS(data_f, "data/table6_2023_Surgeon_nullcpt.rds")
# saveRDS(data_f, paste0("../processed_data/", date_str, "/",
#                        "table6_", year_w, "_Surgeon_nullcpt.rds"))
# glimpse(data_f)
saveRDS(data_f, fp_tab6_Surgeon_nullcpt)

# cat()

# system()
