==> pipeline/6b_check_NA_cols_pred_prob.qmd <==
---
title: "Evaluation of Potentially Missing Predictors"
author: "Aspin Analytical Team"
date: last-modified
execute: 
  warning: false
  echo: false
format:
  html:
    toc: true
    toc-depth: 5
    toc-title: "Jump To"
    highlight-style: github
    code-line-numbers: true
    page-layout: article
    embed-resources: true
editor: source
---

```{r}
#| label: libraries
#| include: false

## set up workspace
rm(list = ls())
library(tidyverse)
library(magrittr)
library(gt)       # html table 
library(knitr)
library(gtsummary)
library(visdat) # missingness pattern
library(bit64)
library(questionr)
library(lubridate) 
library(naniar) # missingness summary

# helper functions
# source("./basic.R")      
# source("./helper.R") 

```


```{r}
#| label: global_vars_data
#| include: false

source("./global_vars.R")


pre_op_missing_cols <- c(
  "OS CT BODY COMPARE-NO READ",
  "ANTIBODY PATIENT INTERPS 1-5",
  "CLOS LARGE BOWEL BIOPSY",
  "UPDATE PATIENT CLASS I.E. INPATIENT / OBSERVATION"
)

post_op_missing_cols <- c(
  "CPAP OVERNIGHT",
  "INSERT ARTERIAL LINE",
  "XR CHEST 2 VIEW (PA,LAT)",
  "POINT OF CARE TESTS",
  "INCISION AND DRAINAGE HEAD/NECK",
  "LAST TYPE & SCREEN (CLOT) STATUS",
  "POCT CRITICAL PANEL",
  "ADMIT PATIENT",
  "FEMORAL NERVE BLOCK",
  "RETURN PATIENT",
  "NSAIDS, CYCLOOXYGENASE INHIBITOR - TYPE ANALGESICS",
  "POCT GLUCOSE CPT 82962 INTERFACED RESULT/DOCKED DEVICE",
  "RESTRAINTS NON-BEHAVIORAL",
  "TRANSFER PATIENT",
  "D DIMER QUANTITATIVE",
  "IP CONSULT TO PHARMACY - TPN",
  "IP CONSULT TO WOUND / OSTOMY / SKIN TEAM",
  "US RENAL (KIDNEYS/BLADDER ONLY)",
  "US UPPER EXTREMITY VENOUS BIL",
  "WEIGH PATIENT",
  "PREADMISSION ADMIT ORDER - RN TO RELEASE",
  "UPDATE PATIENT CLASS I.E. INPATIENT / OBSERVATION",
  ".XR CHEST SINGLE (PA)",
  "CTA CHEST FOR PE",
  "OPN RT HEMICOLECTOMY NEC",
  "US UPPER EXTREMITY VENOUS UNILATERAL"
)


```

# 

## Pre-operative

| Potentially Missing Pre-operative Column Name         | Complication(s) Used In                                  |
| :---------------------------------------------------- | :------------------------------------------------------- |
| `OS CT BODY COMPARE-NO READ`                          | `pred_prob_SSI_preop`, `pred_prob_UTI_preop`, `pred_prob_bleed_preop` |
| `ANTIBODY PATIENT INTERPS 1-5`                        | `pred_prob_cardiac_preop`                                |
| `CLOS LARGE BOWEL BIOPSY`                             | `pred_prob_upradmin_preop`                               |
| `UPDATE PATIENT CLASS I.E. INPATIENT / OBSERVATION`   | `pred_prob_upradmin_preop`                               |


## Post-operative 

| Potentially Missing Post-operative Column Name                | Complication(s) Used In                         |
| :------------------------------------------------------------ | :---------------------------------------------- |
| `CPAP OVERNIGHT`                                              | `pred_prob_cardiac`                             |
| `INSERT ARTERIAL LINE`                                        | `pred_prob_cardiac`                             |
| `XR CHEST 2 VIEW (PA,LAT)`                                    | `pred_prob_cardiac`                             |
| `POINT OF CARE TESTS`                                         | `pred_prob_cardiac`                             |
| `INCISION AND DRAINAGE HEAD/NECK`                             | `pred_prob_morb`                                |
| `LAST TYPE & SCREEN (CLOT) STATUS`                            | `pred_prob_morb`, `pred_prob_bleed`             |
| `POCT CRITICAL PANEL`                                         | `pred_prob_morb`, `pred_prob_bleed`             |
| `ADMIT PATIENT`                                               | `pred_prob_nothome`, `pred_prob_upradmin`       |
| `FEMORAL NERVE BLOCK`                                         | `pred_prob_nothome`                             |
| `RETURN PATIENT`                                              | `pred_prob_nothome`                             |
| `NSAIDS, CYCLOOXYGENASE INHIBITOR - TYPE ANALGESICS`          | `pred_prob_nothome`                             |
| `POCT GLUCOSE CPT 82962 INTERFACED RESULT/DOCKED DEVICE`    | `pred_prob_nothome`                             |
| `RESTRAINTS NON-BEHAVIORAL`                                   | `pred_prob_pulmonary`                           |
| `TRANSFER PATIENT`                                            | `pred_prob_pulmonary`                           |
| `D DIMER QUANTITATIVE`                                        | `pred_prob_pulmonary`                           |
| `IP CONSULT TO PHARMACY - TPN`                                | `pred_prob_renal`                               |
| `IP CONSULT TO WOUND / OSTOMY / SKIN TEAM`                    | `pred_prob_renal`                               |
| `US RENAL (KIDNEYS/BLADDER ONLY)`                             | `pred_prob_renal`                               |
| `US UPPER EXTREMITY VENOUS BIL`                               | `pred_prob_renal`                               |
| `WEIGH PATIENT`                                               | `pred_prob_renal`                               |
| `PREADMISSION ADMIT ORDER - RN TO RELEASE`                    | `pred_prob_upradmin`                            |
| `UPDATE PATIENT CLASS I.E. INPATIENT / OBSERVATION`         | `pred_prob_upradmin`                            |
| `.XR CHEST SINGLE (PA)`                                       | `pred_prob_VTE`                                 |
| `CTA CHEST FOR PE`                                            | `pred_prob_VTE`                                 |
| `OPN RT HEMICOLECTOMY NEC`                                    | `pred_prob_VTE`                                 |
| `US UPPER EXTREMITY VENOUS UNILATERAL`                        | `pred_prob_VTE`                                 |
